==> 02_exposition.Rmd <==
---
title: "2.Exposition: Time weighted average PaO2"
output: html_notebook
---

This notebook computes the time-weighted average PaO2 (TWA-PaO2) over the first 24 hours and selects the final cohort.

```{r}
library(RPostgreSQL)
library(tidyverse)
library(lubridate)

driver <- dbDriver("PostgreSQL")
db <- dbConnect(
  driver,
  dbname = "mimiciii",
  host = "localhost",
  port = "5432",
  user = "mimicuser"
  )
```

```{r}
TimeWeightedAvgPo2 <- function(abg, intime){
  # Compute the time weigted average of several PaO2
  #
  # Args:
  #   abg: Dataframe containing the po2 values and the corresponding charttime
  #   intime: Date and time of subject admission in the ICU
  #
  # Returns:
  #   The time weighted PaO2 mean value in mmHg
  
  abg$weight <- 0

  for (i in 1:nrow(abg)) {
      last <- 0
    
      # First ABG : time difference with intime
      if(i == 1){
        abg[1, "weight"] <- difftime(abg[1, "charttime"], intime, units = "hours")
      }
      else {
        # ABG between the first 24 hours
        if(abg[i, "charttime"] < intime + 24*60*60){
          abg[i, "weight"] <- difftime(abg[i, "charttime"], abg[i-1, "charttime"], units = "hours")
          last <- i
        }
        else {
          # First ABG after the 24 first hours
          if(i == last + 1){
            # Duration until the 24th hour and the last ABG
            abg[i, "weight"] <- difftime(intime+24*60*60, abg[i-1, "charttime"], units = "hours")
          }
        }
      }
  }

  result <- round(weighted.mean(abg$po2, abg$weight), 2)
  result
}
```

# Compute time weighted average PaO2 for every patient

```{r}
cohort_shock <- read.csv("data/cohort_shock.csv")
cohort_shock$intime <- as.POSIXct(cohort_shock$intime)
```

```{r}
query <- "select icustay_id, 
  Cast(charttime as Varchar) as charttime, 
  po2
from abg_sepsis
left join cohort_sepsis as co using (icustay_id)
where co.flag_shock = 1
and charttime >= intime
and po2 is not null
order by icustay_id asc, charttime asc;"
abg_all <- dbGetQuery(db, query)

#abg_all$charttime <- as.POSIXct(strptime(abg_all$charttime, "%Y-%m-%d %H:%M:%S"))
abg_all$charttime <- ymd_hms(abg_all$charttime)
```

```{r}
cohort_shock$twavg_po2 <- 0

for(i in 1:nrow(cohort_shock)){
  id <- cohort_shock[i, "icustay_id"]
  abg <- subset(abg_all, icustay_id == id, select = c("charttime", "po2"))
  intime <- cohort_shock[i, "intime"]
  cohort_shock[i, "twavg_po2"] <- TimeWeightedAvgPo2(abg, intime)
  #print(paste("icustay = ", id, "twavg_po2 = ", TimeWeightedAvgPo2(abg, intime)))
}

glimpse(cohort_shock)
```

# Exposition groups: hypoxia, normoxia, hyperoxia

```{r}
cohort_shock$hypero2 <- cut(cohort_shock$twavg_po2, breaks = c(0, 70, 120, Inf), include.lowest = TRUE)
levels(cohort_shock$hypero2) <- c("hypo", "normo", "hyper")
```

```{r}
table(cohort_shock$hypero2)
```

# Dataset

```{r}
cohort_shock$X <- NULL
cohort_shock$first_careunit <- factor(cohort_shock$first_careunit)
cohort_shock$first_wardid <- factor(cohort_shock$first_wardid)
cohort_shock$admission_type <- factor(cohort_shock$admission_type)
cohort_shock$vaso_firstday <- factor(cohort_shock$vaso_firstday)
cohort_shock$gender <- factor(cohort_shock$gender)
cohort_shock$rrt <- factor(cohort_shock$rrt)
cohort_shock$comorbidity_cardio <- factor(cohort_shock$comorbidity_cardio)
cohort_shock$comorbidity_vascular <- factor(cohort_shock$comorbidity_vascular)
cohort_shock$comorbidity_neuro <- factor(cohort_shock$comorbidity_neuro)
cohort_shock$comorbidity_pneumo <- factor(cohort_shock$comorbidity_pneumo)
cohort_shock$comorbidity_metabolic <- factor(cohort_shock$comorbidity_metabolic)
cohort_shock$comorbidity_renal <- factor(cohort_shock$comorbidity_renal)
cohort_shock$comorbidity_liver <- factor(cohort_shock$comorbidity_liver)
cohort_shock$comorbidity_cancer <- factor(cohort_shock$comorbidity_cancer)
cohort_shock$noso_vap <- factor(cohort_shock$noso_vap)
cohort_shock$noso_urinary <- factor(cohort_shock$noso_urinary)
cohort_shock$noso_cvc <- factor(cohort_shock$noso_cvc)
cohort_shock$noso_cvc_other <- factor(cohort_shock$noso_cvc_other)
cohort_shock$source_pulm <- factor(cohort_shock$source_pulm)
cohort_shock$source_urogyn <- factor(cohort_shock$source_urogyn)
cohort_shock$source_neuro <- factor(cohort_shock$source_neuro)
cohort_shock$source_dig <- factor(cohort_shock$source_dig)
cohort_shock$source_ost <- factor(cohort_shock$source_ost)
cohort_shock$source_ent <- factor(cohort_shock$source_ent)
cohort_shock$source_card <- factor(cohort_shock$source_card)
cohort_shock$source_skin <- factor(cohort_shock$source_skin)
cohort_shock$source_other <- factor(cohort_shock$source_other)
cohort_shock$death_icu <- factor(cohort_shock$death_icu)
cohort_shock$death_thirtydays <- factor(cohort_shock$death_thirtydays)
cohort_shock$death_hosp <- factor(cohort_shock$death_hosp)
cohort_shock$flag_shock <- factor(cohort_shock$flag_shock)
cohort_shock$hypero2 <- factor(cohort_shock$hypero2)
cohort_shock$hypero2 <- relevel(cohort_shock$hypero2, ref = "normo")

# Merge Emergency and Urgent admission types
cohort_shock$admission_type <- as.character(cohort_shock$admission_type)
cohort_shock$admission_type[cohort_shock$admission_type %in% c("EMERGENCY", "URGENT")] <- "Emergency/Urgent"
cohort_shock$admission_type <- factor(cohort_shock$admission_type)
levels(cohort_shock$admission_type) <- c("Elective", "Emergency/Urgent")

saveRDS(cohort_shock, "data/cohort_shock_twavgpo2_all.rds")
```

```{r}
# Exclude hypoxemia patients
cohort_shock <- cohort_shock %>% 
  filter(hypero2 != "hypo")

cohort_shock$hypero2 <- factor(cohort_shock$hypero2)
```

```{r}
saveRDS(cohort_shock, "data/cohort_shock_twavgpo2.rds")
```

```{r}
dbDisconnect(db)
dbUnloadDriver(driver)
```
