==> 04_eda.Rmd <==
---
title: "4.Exploratory data analysis"
output:
  html_document:
    df_print: paged
---

```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(knitr)
library(kableExtra)
library(reshape2)

cohort <- readRDS("data/cohort_shock_twavgpo2.rds")
```

```{r}
summary(cohort)
```

Replace anonymized ages (above 300 y/o) with 89

```{r}
cohort <- cohort %>%
  mutate(admission_age = replace(admission_age, admission_age > 300, 89))
```

# Continuous variables

```{r}
cohort %>%
  keep(is.numeric) %>%
  select(-subject_id, -hadm_id, -icustay_id) %>%
  map_df(.f = ~ broom::tidy(round(summary(.x), 2)), .id = "variable") %>%
  kable(caption = "Continuous variables",
        col.names = c("Variables", "Min", "25 %", "Median", "Mean", "75 %", "Max", "NA")) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))
```

```{r message=FALSE, warning=FALSE, dpi=400}
cohort %>%
  keep(is.numeric) %>%
  select(-subject_id, -hadm_id, -icustay_id) %>%
  melt() %>%
  ggplot(aes(x = value)) +
  facet_wrap(~ variable, scales = "free", ncol = 5) +
  geom_histogram(fill = "#455a64") +
  labs(x = "Continuous variables", y ="")
```

# Categorial variables

```{r message=FALSE, warning=FALSE}
cohort %>%
  keep(is.factor) %>%
  gather(var, value, everything()) %>%
  group_by(var, value) %>%
  summarise(n = n()) %>%
  mutate(freq = round((n / sum(n))*100, 2)) %>%
  kable(caption = "Categorial variables",
        col.names = c("Variables", "Value", "Count", "Proportion (%)")) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))
```

```{r message=FALSE, warning=FALSE, dpi=400}
cohort %>%
  keep(is.factor) %>%
  gather() %>%
  ggplot(aes(x = value)) +
  facet_wrap(~ key, scales = "free") +
  geom_bar(fill = "#455a64") +
  labs(x = "Categorical variables", y ="")
```

# Missing values

```{r}
cohort %>%
  summarise_all(list(~sum(is.na(.)))) %>%
  gather(variable, nb_na) %>%
  arrange(desc(nb_na)) %>%
  kable(caption = "Missing values",
        col.names = c("Variables", "Missing values")) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))
```

```{r dpi=400}
cohort %>%
  rename(
    "BMI"  = bmi_firstday,
    "Height" = height_firstday,
    "Weight" = weight_firstday,
    "Albumin" = albumin_min_firstday,
    "Bilirubin" = bilirubin_max_firstday,
    "PTT" = ptt_min_firstday
  ) %>%
  summarise_all(list(~sum(is.na(.)))) %>%
  gather(variable, nb_na) %>%
  filter(nb_na > 0) %>%
  ggplot(aes(x = reorder(variable, nb_na), y = nb_na)) +
  geom_bar(stat = "identity") +
  geom_text(aes(label = nb_na), hjust = -0.2, size = 8) +
  coord_flip() +
  labs(
    y = "Missing values",
    x = "Variables"
  ) +
  theme_bw() +
  theme(
    axis.title.x = element_text(size = 24),
    axis.title.y = element_text(size = 24),
    axis.text.x = element_text(size = 18),
    axis.text.y = element_text(size = 18)
  )
```

```{r}
cohort %>%
  group_by(twavg_po2_interv = cut_interval(twavg_po2, length = 10)) %>%
  count(death_icu) %>% 
  pivot_wider(names_from = death_icu, values_from = n, names_prefix = "death_icu_") %>% 
  mutate(death_icu_prob = death_icu_1/(death_icu_0 + death_icu_1)) %>% 
  ggplot(aes(x = twavg_po2_interv, y = death_icu_prob)) +
  geom_bar(stat = "identity") +
  theme(
    axis.text.x = element_text(angle = 45, vjust = 0.5)
  ) +
  labs(
    y = "In-ICU death ratio",
    x = "TWA-PaO2"
  )
```
