==> 05_outcomes.Rmd <==
---
title: "5.Outcomes"
output:
  html_document:
    df_print: paged
---

```{r message=FALSE, warning=FALSE, include=FALSE}
library(tidyverse)
library(tableone)
library(knitr)
library(kableExtra)
library(Epi)
library(survival)
library(broom)
library(corrplot)
library(patchwork)
library(survminer)
library(MBESS)
library(VGAM)

palette <- c("#0099CC", "#D55E00")

cohort <- readRDS("data/cohort_shock_twavgpo2.rds")

# Replace anonymized ages (above 300 y/o) with 89
cohort <- cohort %>%
  mutate(admission_age = replace(admission_age, admission_age > 300, 89))
```

# Baseline demographics

```{r message=FALSE, warning=FALSE}
table_demo_data <- cohort %>% 
  select(
    admission_age,
    gender,
    bmi_firstday,
    sapsii,
    sofa,
    first_careunit,
    admission_type,
    comorbidity_cardio,
    comorbidity_vascular,
    comorbidity_neuro,
    comorbidity_pneumo,
    comorbidity_metabolic,
    comorbidity_renal,
    comorbidity_liver,
    comorbidity_cancer,
    hypero2
  ) %>%
  rename(
    "Age (years)" = admission_age,
    "Sexe (Male)" = gender,
    "BMI (kg/m2)" = bmi_firstday,
    "SAPSII" = sapsii,
    "SOFA" = sofa,
    "First ICU (SICU)" = first_careunit,
    "Admission type (Emergency/Urgent)" = admission_type,
    "Cardiac" = comorbidity_cardio,
    "Vascular" = comorbidity_vascular,
    "Neurological" = comorbidity_neuro,
    "Pulmonary" = comorbidity_pneumo,
    "Metabolic" = comorbidity_metabolic,
    "Renal" = comorbidity_renal,
    "Liver" = comorbidity_liver,
    "Cancer" = comorbidity_cancer
  )

factor_vars <- c(
  "Sexe (Male)", "First ICU (SICU)", "Admission type (Emergency/Urgent)",
  "Cardiac", "Vascular", "Neurological", "Pulmonary", "Metabolic",
  "Renal", "Liver", "Cancer"
  )
nonnormal_vars <- c(
  "Age (years)", "BMI (kg/m2)", "SAPSII", "SOFA"
  )

table_demo <- CreateTableOne(
  vars = names(table_demo_data)[-length(names(table_demo_data))],
  data = table_demo_data,
  factorVars = factor_vars,
  strata = "hypero2"
)
table_demo_print <- print(
  table_demo,
  nonnormal = nonnormal_vars,
  dropEqual = TRUE,
  printToggle = FALSE
)
```

```{r}
table_demo_all <- CreateTableOne(
  vars = names(table_demo_data)[-length(names(table_demo_data))],
  data = table_demo_data,
  factorVars = factor_vars
)
table_demo_all_print <- print(
  table_demo_all,
  nonnormal = nonnormal_vars,
  dropEqual = TRUE,
  printToggle = FALSE
)
```

```{r}
table_demo_final_print <- cbind(table_demo_all_print[, 1], table_demo_print[, c(1:3)])

kable(
  table_demo_final_print,
  caption = "Baseline demographics of the cohort for all patients as well as normoxemia and hyperoxemia groups",
  col.names = c("All", "Normoxia", "Hyperoxia", "p"),
  align = c("l", "c", "c", "c", "c")
) %>%
  kable_styling(bootstrap_options = c("striped", "hover")) %>%
  group_rows("Comorbidities", 9, 16)
```

# Infection sources

```{r}
table_sources_data <- cohort %>%
  select(
    source_pulm,
    source_urogyn,
    source_neuro,
    source_dig,
    source_ost,
    source_ent,
    source_card,
    source_skin,
    source_other,
    hypero2
  ) %>%
  rename(
    "Pulmonary" = source_pulm,
    "Neurological" = source_neuro,
    "Genito-urinary" = source_urogyn,
    "Digestive" = source_dig,
    "Osteo-articular" = source_ost,
    "ENT" = source_ent,
    "Cardiological" = source_card,
    "Cutaneus" = source_skin,
    "Other/Unknown" = source_other
  )
factor_vars <- c("Pulmonary", "Neurological", "Genito-urinary", "Digestive",
                 "Osteo-articular", "ENT", "Cardiological", "Cutaneus",
                 "Other/Unknown")

table_sources <- CreateTableOne(
  vars = names(table_sources_data)[-length(names(table_sources_data))],
  data = table_sources_data,
  factorVars = factor_vars,
  strata = "hypero2"
)
table_sources_print <- print(
  table_sources,
  dropEqual = TRUE,
  printToggle = FALSE
)
```

```{r}
table_sources_all <- CreateTableOne(
  vars = names(table_sources_data)[-length(names(table_sources_data))],
  data = table_sources_data,
  factorVars = factor_vars
)
table_sources_all_print <- print(
  table_sources_all,
  dropEqual = TRUE,
  printToggle = FALSE
)
```

```{r}
table_sources_final_print <- cbind(table_sources_all_print[, 1], table_sources_print[, c(1:3)])

kable(
  table_sources_final_print,
  caption = "Infection sources",
  col.names = c("All", "Normoxia", "Hyperoxia", "p"),
  align = c("l", "c", "c", "c", "c")
) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))
```

# Biology results

```{r}
table_bio_data <- cohort %>%
  select(
    twavg_po2,
    first_po2,
    avg_24_po2,
    min_24_po2,
    max_24_po2,
    lactate_max_firstday,
    pco2_max_firstday,
    ph_min_firstday,
    bicarbonate_min_firstday,
    albumin_min_firstday,
    bilirubin_max_firstday,
    creatinine_max_firstday,
    chloride_max_firstday,
    sodium_max_firstday,
    potassium_max_firstday,
    hematocrit_min_firstday,
    hemoglobin_min_firstday,
    platelet_min_firstday,
    ptt_min_firstday,
    hypero2
  ) %>%
  rename(
    "TWA-PaO2 (mmHg)" = twavg_po2,
    "First PaO2 (mmHg)" = first_po2,
    "Average PaO2 (mmHg)" = avg_24_po2,
    "Lowest PaO2 (mmHg)" = min_24_po2,
    "Highest PaO2 (mmHg)" = max_24_po2,
    "Serum lactate (mmol/L)" = lactate_max_firstday,
    "PaCO2 (mmHg)" = pco2_max_firstday,
    "pH" = ph_min_firstday,
    "Bicarbonate (mmol/L)" = bicarbonate_min_firstday,
    "Albumin (g/dL)" = albumin_min_firstday,
    "Bilirubin (mg/dL)" = bilirubin_max_firstday,
    "Creatinine (mg/dL)" = creatinine_max_firstday,
    "Chloride (mmol/L)" = chloride_max_firstday,
    "Sodium (mmol/L)" = sodium_max_firstday,
    "Potassium (mmol/L)" = potassium_max_firstday,
    "Hematocrit (%)" = hematocrit_min_firstday,
    "Hemoglobine (g/dL)" = hemoglobin_min_firstday,
    "Platelets (G/L)" = platelet_min_firstday,
    "PTT (%)" = ptt_min_firstday
  )

table_bio <- CreateTableOne(
  vars = names(table_bio_data)[-length(names(table_bio_data))],
  data = table_bio_data,
  strata = c("hypero2")
)
table_bio_print <- print(
  table_bio,
  nonnormal = names(table_bio_data)[-length(names(table_bio_data))],
  dropEqual = TRUE,
  printToggle = FALSE
)
```

```{r}
table_bio_all <- CreateTableOne(
  vars = names(table_bio_data)[-length(names(table_bio_data))],
  data = table_bio_data
)
table_bio_all_print <- print(
  table_bio_all,
  nonnormal = names(table_bio_data)[-length(names(table_bio_data))],
  dropEqual = TRUE,
  printToggle = FALSE
)
```

```{r echo=FALSE}
table_bio_final_print <- cbind(table_bio_all_print[, 1], table_bio_print[, c(1:3)])

kable(
  table_bio_final_print,
  caption = "Biology results of the first 24 hours",
  col.names = c("All", "Normoxia", "Hyperoxia", "p"),
  align = c("l", "c", "c", "c", "c")
) %>%
  kable_styling(bootstrap_options = c("striped", "hover")) %>%
  group_rows("Arterial blood gases", 2, 9) %>%
  group_rows("Chemistry", 10, 16) %>%
  group_rows("Blood count and hemostasis", 17, 20)
```

## Survivors vs non survivors

```{r message=FALSE, warning=FALSE}
table_surv_data <- cohort %>% 
  select(
    admission_age,
    gender,
    bmi_firstday,
    sapsii,
    sofa,
    first_careunit,
    admission_type,
    comorbidity_cardio,
    comorbidity_vascular,
    comorbidity_neuro,
    comorbidity_pneumo,
    comorbidity_metabolic,
    comorbidity_renal,
    comorbidity_liver,
    comorbidity_cancer,
    twavg_po2,
    hypero2,
    lactate_max_firstday,
    pco2_max_firstday,
    ph_min_firstday,
    bicarbonate_min_firstday,
    albumin_min_firstday,
    bilirubin_max_firstday,
    creatinine_max_firstday,
    hemoglobin_min_firstday,
    source_pulm,
    source_neuro,
    source_urogyn,
    source_dig,
    source_ost,
    source_ent,
    source_card,
    source_skin,
    source_other,
    norepinephrine_max_dose,
    norepinephrine_duration,
    death_icu
  ) %>%
  rename(
    "Age (years)" = admission_age,
    "Sex (Male)" = gender,
    "BMI (kg/m2)" = bmi_firstday,
    "SAPSII" = sapsii,
    "SOFA" = sofa,
    "First ICU (SICU)" = first_careunit,
    "Admission type (Emergency/Urgent)" = admission_type,
    "Cardiac" = comorbidity_cardio,
    "Vascular" = comorbidity_vascular,
    "Neurological" = comorbidity_neuro,
    "Pulmonary" = comorbidity_pneumo,
    "Metabolic" = comorbidity_metabolic,
    "Renal" = comorbidity_renal,
    "Liver" = comorbidity_liver,
    "Cancer" = comorbidity_cancer,
    "TWA-PaO2 (mmHg)" = twavg_po2,
    "Serum lactate (mmol/L)" = lactate_max_firstday,
    "PaCO2 (mmHg)" = pco2_max_firstday,
    "pH" = ph_min_firstday,
    "Bicarbonate (mmol/L)" = bicarbonate_min_firstday,
    "Albumin (g/dL)" = albumin_min_firstday,
    "Bilirubin (mg/dL)" = bilirubin_max_firstday,
    "Creatinine (mg/dL)" = creatinine_max_firstday,
    "Hemoglobin (g/dL)" = hemoglobin_min_firstday,
    "Hyperoxemia (TWA-PaO2 > 120 mmHg)" = hypero2
  )

factor_vars <- c(
  "Sex (Male)", "First ICU (SICU)", "Admission type (Emergency/Urgent)",
  "Cardiac", "Vascular", "Neurological", "Pulmonary", "Metabolic",
  "Renal", "Liver", "Cancer", "Hyperoxemia (TWA-PaO2 > 120 mmHg)",
  "source_pulm", "source_neuro", "source_urogyn", "source_dig",
  "source_ost", "source_ent", "source_card", "source_skin", "source_other"
  )
nonnormal_vars <- c(
  "Age (years)", "BMI (kg/m2)", "SAPSII", "SOFA", "TWA-PaO2 (mmHg)",
  "Serum lactate (mmol/L)", "PaCO2 (mmHg)", "pH", "Bicarbonate (mmol/L)",
  "Albumin (g/dL)", "Bilirubin (mg/dL)", "Creatinine (mg/dL)", "Hemoglobin (g/dL)"
  )

table_surv <- CreateTableOne(
  vars = names(table_surv_data)[-length(names(table_surv_data))],
  data = table_surv_data,
  factorVars = factor_vars,
  strata = "death_icu"
)
table_surv_print <- print(
  table_surv,
  nonnormal = nonnormal_vars,
  dropEqual = TRUE,
  printToggle = FALSE
)
```

```{r}
table_surv_all <- CreateTableOne(
  vars = names(table_surv_data)[-length(names(table_surv_data))],
  data = table_surv_data,
  factorVars = factor_vars
)
table_surv_all_print <- print(
  table_surv_all,
  nonnormal = nonnormal_vars,
  dropEqual = TRUE,
  printToggle = FALSE
)
```

```{r}
table_surv_final_print <- cbind(table_surv_all_print[, 1], table_surv_print[, c(1:3)])

kable(
  table_surv_final_print,
  caption = "Baseline demographics of the cohort for all patients as well as normoxemia and hyperoxemia groups",
  col.names = c("All", "Survived", "Deceased", "p"),
  align = c("l", "c", "c", "c", "c")
) %>%
  kable_styling(bootstrap_options = c("striped", "hover")) %>%
  group_rows("Comorbidities", 9, 16) %>% 
  group_rows("Biology", 17, 26) %>% 
  group_rows("Infection sources", 27, 35)
```

# Univariate analyses

```{r}
table_bivariate_data <- cohort %>%
  select(
    death_icu,
    death_hosp,
    death_thirtydays,
    los_icu,
    los_hospital,
    norepinephrine_max_dose,
    norepinephrine_duration,
    rrt,
    noso_vap,
    noso_urinary,
    noso_cvc_other,
    hypero2,
  ) %>%
  rename(
    "ICU LOS" = los_icu,
    "Hospital LOS" = los_hospital,
    "Norepinephrine maximum dose (mcg/kg/min)" = norepinephrine_max_dose,
    "Norepinephrine duration (hours)" = norepinephrine_duration,
    "Ventilation acquired pneumoniae" = noso_vap,
    "Urinary catheter infection" = noso_urinary,
    "Central line infection" = noso_cvc_other,
    "ICU mortality" = death_icu,
    "Hospital mortality" = death_hosp,
    "30-day mortality" = death_thirtydays
  )

vars <- c(
  "ICU mortality", "Hospital mortality", "30-day mortality", "ICU LOS",
  "Hospital LOS", "Norepinephrine maximum dose (mcg/kg/min)",
  "Norepinephrine duration (hours)", "Ventilation acquired pneumoniae",
  "Urinary catheter infection", "Central line infection"
  )
factor_vars <- c(
  "Ventilation acquired pneumoniae", "Urinary catheter infection",
  "Central line infection", "30-day mortality", "ICU mortality",
  "Hospital mortality"
  )
nonnormal_vars <- c(
  "ICU LOS", "Hospital LOS", "Norepinephrine maximum dose (mcg/kg/min)",
  "Norepinephrine duration (hours)"
  )

table_bivar <- CreateTableOne(
  vars = vars,
  data = table_bivariate_data,
  strata = c("hypero2")
)
table_bivar_print <- print(
  table_bivar,
  nonnormal = nonnormal_vars,
  dropEqual = TRUE,
  printToggle = FALSE,
  smd = TRUE
)
```

```{r}
# bivar_smd <- ExtractSmd(table_bivar)
# smd_list <- tibble(
#   smd = as.character(),
#   low = as.character(),
#   high = as.character(),
#   txt = as.character())
# smd_list <- smd_list %>% add_row(smd = "", low = "", high = "", txt = "")
# 
# for(i in 1:length(bivar_smd)){
#   smd_ci <- ci.smd(smd = bivar_smd[i], n.1 = 214, n.2 = 274, conf.level = .95)
#   smd_list <- smd_list %>%
#     add_row(
#       smd = round(bivar_smd[i], 2),
#       low = round(smd_ci$Lower.Conf.Limit.smd, 2),
#       high = round(smd_ci$Upper.Conf.Limit.smd, 2))
# }
# 
# smd_list$txt <- paste0(smd_list$smd, " [", smd_list$low, "-", smd_list$high, "]")
# smd_list[1,4] <- ""
```

```{r}
table_bivar_final_print <- cbind(table_bivar_print[, c(1:2)], table_bivar_print[, 3])
kable(
  table_bivar_final_print,
  caption = "Primary and secondary outcomes",
  col.names = c("Normoxia", "Hyperoxia", "p"),
  align = c("l", "c", "c", "c")
) %>%
  kable_styling(bootstrap_options = c("striped", "hover")) %>%
  group_rows("Primary outcome", 2, 2) %>%
  group_rows("Mortality", 3, 4) %>%
  group_rows("Length of stay", 5, 6) %>%
  group_rows("Vasopressor", 7, 8) %>%
  group_rows("Hospital acquired infections", 9, 11)
```

## Unadjusted odds ratios

```{r}
cohort$hypero2_bin <- ifelse(cohort$hypero2 == "hyper", 1, 0)
table(cohort$hypero2, cohort$hypero2_bin)
```

```{r include=FALSE}
# Death ICU
or_death_icu <- twoby2(1 - as.numeric(cohort$hypero2_bin), 1 - as.numeric(cohort$death_icu))
list_crude_or <- round(or_death_icu$measures[2, 1], 2)
list_crude_low <- round(or_death_icu$measures[2, 2], 2)
list_crude_high <- round(or_death_icu$measures[2, 3], 2)

# Death hosp
or_death_hosp <- twoby2(1 - as.numeric(cohort$hypero2_bin), 1 - as.numeric(cohort$death_hosp))
list_crude_or <- c(list_crude_or,
                   round(or_death_hosp$measures[2, 1], 2))
list_crude_low <- c(list_crude_low, round(or_death_hosp$measures[2, 2], 2))
list_crude_high <- c(list_crude_high, round(or_death_hosp$measures[2, 3], 2))

# Death 30 days
or_death_thirty <- twoby2(1 - as.numeric(cohort$hypero2_bin), 1 - as.numeric(cohort$death_thirtydays))
list_crude_or <- c(list_crude_or,
                   round(or_death_thirty$measures[2, 1], 2))
list_crude_low <- c(list_crude_low, round(or_death_thirty$measures[2, 2], 2))
list_crude_high <- c(list_crude_high, round(or_death_thirty$measures[2, 3], 2))

# ICU LOS
mod_or_los_icu <- glm(hypero2_bin ~ los_icu, family = "binomial", data = cohort)
list_crude_or <- c(list_crude_or,
                   round(exp(tidy(mod_or_los_icu)$estimate[2]), 2))
list_crude_low <- c(list_crude_low, round(exp(confint_tidy(mod_or_los_icu)), 2)$conf.low[2])
list_crude_high <- c(list_crude_high, round(exp(confint_tidy(mod_or_los_icu)), 2)$conf.high[2])

# Hospital LOS
mod_or_los_hosp <- glm(hypero2_bin ~ los_hospital, family = "binomial", data = cohort)
list_crude_or <- c(list_crude_or,
                   round(exp(tidy(mod_or_los_hosp)$estimate[2]), 2))
list_crude_low <- c(list_crude_low, round(exp(confint_tidy(mod_or_los_hosp)), 2)$conf.low[2])
list_crude_high <- c(list_crude_high, round(exp(confint_tidy(mod_or_los_hosp)), 2)$conf.high[2])

tab_crude_or <- tibble(
  label = c("death_icu", "death_hosp", "death_thirtydays", "los_icu", "los_hospital"),
  or = list_crude_or,
  low = list_crude_low,
  high = list_crude_high,
)
tab_crude_or$IC_95 <- paste0("[", tab_crude_or$low, "-", tab_crude_or$high, "]")
```

```{r echo=FALSE}
tab_crude_or %>%
  select(label, or, IC_95) %>%
  kable(
    caption = "Unadjusted odds ratios",
    col.names = c("", "OR", "CI (95%)"),
    align = c("c", "c", "c", "c")
  ) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))
```

```{r echo=FALSE}
tab_crude_or$label <- factor(tab_crude_or$label, levels = rev(tab_crude_or$label))

ggplot(tab_crude_or, aes(x = label, y = or, ymin = low, ymax = high)) +
  geom_pointrange(color = "#455a64") +
  geom_hline(yintercept = 1, lty = 2, color = "#455a64") + 
  coord_flip() +
  labs(x = "", y = "Unadjusted OR (95% CI)") +
  theme_bw()
```

## TWA-PaO2 distribution

```{r message=FALSE, warning=FALSE, dpi=400}
ggplot(cohort) +
  geom_histogram(aes(x = twavg_po2, fill = hypero2), alpha = 0.8, bins = 75) +
  labs(
    x = "TWA-PaO2 (mmHg)",
    y = "Count"
  ) +
  # scale_fill_grey(
  #   name = "Groups",
  #   labels = c("Normoxia", "Hyperoxia")
  # ) +
  theme_bw()
  # theme(
  #   legend.position = "bottom",
  #   axis.title.x = element_text(size = 24),
  #   axis.title.y = element_text(size = 24),
  #   axis.text.x = element_text(size = 18),
  #   axis.text.y = element_text(size = 18),
  #   legend.title = element_text(size = 18),
  #   legend.text = element_text(size = 18)
  # )
```




# Multivariate analysis

```{r}
data_corr <- cohort %>%
  select(
    sapsii, sofa, comorbidity_liver, comorbidity_cancer, hypero2_bin,
    lactate_max_firstday, ph_min_firstday, bicarbonate_min_firstday,
    bilirubin_max_firstday, creatinine_max_firstday, source_other
  ) %>% 
  mutate_all(as.numeric)

corrplot(cor(data_corr, use = "complete.obs"), method = "circle", tl.cex = 0.6)
```

## ICU mortality

```{r}
model_icu <- glm(
  death_icu ~ hypero2_bin + sofa + comorbidity_liver + comorbidity_cancer +
    lactate_max_firstday + ph_min_firstday + bicarbonate_min_firstday +
    bilirubin_max_firstday + creatinine_max_firstday + source_other,
  data = cohort,
  family = "binomial"
)
model_icu_desc <- tidy(model_icu)

summary(model_icu)
```

```{r}
tab_reglog <- tibble(
  label = model_icu_desc$term,
  OR = round(exp(model_icu_desc$estimate), 2),
  low = round(exp(confint_tidy(model_icu)), 2)$conf.low,
  high = round(exp(confint_tidy(model_icu)), 2)$conf.high,
  p = round(model_icu_desc$p.value, 3)
) %>% 
  mutate(
    label = case_when(
      label == "hypero2_bin" ~ "Hyperoxemia",
      label == "sofa" ~ "SOFA score",
      label == "comorbidity_liver1" ~ "Hepatic comorbidity",
      label == "comorbidity_cancer1" ~ "Cancer",
      label == "lactate_max_firstday" ~ "Serum lactate",
      label == "ph_min_firstday" ~ "pH",
      label == "bicarbonate_min_firstday" ~ "Bicarbonate",
      label == "bilirubin_max_firstday" ~ "Bilirubin",
      label == "creatinine_max_firstday" ~ "Creatinine",
      label == "source_other1" ~ "Other/unknown infection source"
    ),
    IC_95 = paste0("[", low, "-", high, "]")
  )

ktab_reglog <- tab_reglog %>%
  filter(label != "(Intercept)") %>% 
  select(label, OR, IC_95, p) %>%
  kable(
    col.names = c("", "OR", "95% CI", "p"),
    align = c("c", "c", "c", "c")
  ) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))

ktab_reglog
```

```{r}
reglog_plot <- tab_reglog %>% 
  mutate(label = factor(label, levels = rev(label))) %>% 
  filter(label != "(Intercept)") %>% 
  ggplot(aes(x = label, y = OR, ymin = low, ymax = high)) +
  geom_pointrange(color = "#455a64") +
  geom_hline(yintercept = 1, lty = 2, color = "#455a64") + 
  coord_flip() +
  labs(x = "", y = "Adjusted OR (95% Confidence Interval)") +
  theme_bw()

reglog_plot
```
```{r eval=FALSE, include=FALSE}
tab_reglog <- tab_reglog %>% 
  select(label, OR, IC_95, p) %>% 
  mutate(p = as.character(p)) %>% 
  filter(!is.na(label)) %>% 
  rename(
    "Variable" = label,
    "95% CI" = IC_95,
    "p value" = p
  )

tab_reglog[4, 4] <- "<0.001"

reglog_plot_print <- reglog_plot / gridExtra::tableGrob(tab_reglog, rows = NULL)

ggsave(
  filename = "sfigure2_reglog.eps",
  plot = print(reglog_plot_print),
  width = 8, height = 7, dpi = 600)
```

## Hospital mortality

```{r}
model_hosp <- glm(
  death_hosp ~ hypero2_bin + sofa + comorbidity_liver + comorbidity_cancer +
    lactate_max_firstday + ph_min_firstday + bicarbonate_min_firstday +
    bilirubin_max_firstday + creatinine_max_firstday + source_other,
  data = cohort,
  family = "binomial"
)
model_hosp_desc <- tidy(model_hosp)

summary(model_hosp)
```

```{r}
tibble(
  label = model_hosp_desc$term,
  OR = round(exp(model_hosp_desc$estimate), 2),
  low = round(exp(confint_tidy(model_hosp)), 2)$conf.low,
  high = round(exp(confint_tidy(model_hosp)), 2)$conf.high,
  p = round(model_hosp_desc$p.value, 5)
) %>% 
  mutate(IC_95 = paste0("[", low, "-", high, "]")) %>% 
  select(label, OR, IC_95, p) %>%
  kable(
    caption = "Multivariate analysis of hospital mortality (logistic regression)",
    col.names = c("", "OR", "IC (95%)", "p"),
    align = c("c", "c", "c", "c")
  ) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))
```

## 30-days mortality

```{r}
model_thirtydays <- glm(
  death_thirtydays ~ hypero2_bin + sofa + comorbidity_liver + comorbidity_cancer +
    lactate_max_firstday + ph_min_firstday + bicarbonate_min_firstday +
    bilirubin_max_firstday + creatinine_max_firstday + source_other,
  data = cohort,
  family = "binomial"
)
model_thirtydays_desc <- tidy(model_thirtydays)

summary(model_thirtydays)
```

```{r}
tibble(
  label = model_thirtydays_desc$term,
  OR = round(exp(model_thirtydays_desc$estimate), 2),
  low = round(exp(confint_tidy(model_thirtydays)), 2)$conf.low,
  high = round(exp(confint_tidy(model_thirtydays)), 2)$conf.high,
  p = round(model_thirtydays_desc$p.value, 5)
) %>% 
  mutate(IC_95 = paste0("[", low, "-", high, "]")) %>% 
  select(label, OR, IC_95, p) %>%
  kable(
    caption = "Multivariate analysis of 30-days mortality (logistic regression)",
    col.names = c("", "OR", "IC (95%)", "p"),
    align = c("c", "c", "c", "c")
  ) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))
```

## ICU length of stay

```{r}
model_los <- vglm(
  as.integer(los_icu) ~ hypero2_bin + sofa + comorbidity_liver + comorbidity_cancer +
    lactate_max_firstday + ph_min_firstday + bicarbonate_min_firstday +
    bilirubin_max_firstday + creatinine_max_firstday + source_other,
  data = cohort,
  family = posnegbinomial()
)
model_los_sum <- summary(model_los)
model_los_sum 
```

```{r}
tibble(
  label = names(model_los@coefficients),
  OR = round(exp(model_los@coefficients), 2),
  low = round(exp(confint(model_los)[, "2.5 %"]), 2),
  high = round(exp(confint(model_los)[, "97.5 %"]), 2),
  p = round(model_los_sum@coef3[, "Pr(>|z|)"], 3)
) %>% 
  mutate(IC_95 = paste0("[", low, "-", high, "]")) %>% 
  select(label, OR, IC_95, p) %>%
  kable(
    caption = "Multivariate analysis of ICU LOS (zero truncated negative binomial regression)",
    col.names = c("", "OR", "IC (95%)", "p"),
    align = c("c", "c", "c", "c")
  ) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))
```

## Hospital length of stay

```{r}
model_los_hosp <- vglm(
  as.integer(los_hospital) ~ hypero2_bin + sofa + comorbidity_liver + comorbidity_cancer +
    lactate_max_firstday + ph_min_firstday + bicarbonate_min_firstday +
    bilirubin_max_firstday + creatinine_max_firstday + source_other,
  data = cohort,
  family = posnegbinomial()
)
model_los_hosp_sum <- summary(model_los_hosp)
model_los_hosp_sum 
```

```{r}
tibble(
  label = names(model_los_hosp@coefficients),
  OR = round(exp(model_los_hosp@coefficients), 2),
  low = round(exp(confint(model_los_hosp)[, "2.5 %"]), 2),
  high = round(exp(confint(model_los_hosp)[, "97.5 %"]), 2),
  p = round(model_los_hosp_sum@coef3[, "Pr(>|z|)"], 3)
) %>% 
  mutate(IC_95 = paste0("[", low, "-", high, "]")) %>% 
  select(label, OR, IC_95, p) %>%
  kable(
    caption = "Multivariate analysis of hospital LOS (zero truncated negative binomial regression)",
    col.names = c("", "OR", "IC (95%)", "p"),
    align = c("c", "c", "c", "c")
  ) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))
```

## Adjusted probability of in-ICU death by TWA-PaO2

```{r message=FALSE, warning=FALSE, dpi=400}
data_plot_death <- tibble(
  prob = model_icu$fitted.values,
  twavg_po2 = cohort[as.numeric(names(model_icu$fitted.values)), 'twavg_po2']
)

ggplot(data_plot_death, aes(x = twavg_po2, y = prob)) +
  geom_point() +
  geom_smooth(method = loess) +
  labs(
    y = "Adjusted probability of in-ICU death",
    x = "Time weighted average PaO2 (mmHg)"
  ) +
  scale_y_continuous(limits = c(0, 1)) +
  theme_bw()
```

# Survival analysis

```{r}
fit <- survfit(Surv(time = los_icu, event = as.numeric(death_icu))~hypero2, data = cohort)
```

## Log-Rank test

```{r}
survdiff(Surv(time = los_icu, event = as.numeric(death_icu))~hypero2, data = cohort)
```

## Cox model

```{r}
model_cox <- coxph(Surv(time = los_icu, event = as.numeric(death_icu))~hypero2_bin + sofa + comorbidity_liver + comorbidity_cancer +
    lactate_max_firstday + ph_min_firstday + bicarbonate_min_firstday +
    bilirubin_max_firstday + creatinine_max_firstday + source_other, data = cohort)
summary(model_cox)
model_cox_desc <- tidy(model_cox, conf.int = TRUE)
```

```{r}
tab_cox <- tibble(
  label = model_cox_desc$term,
  OR = round(exp(model_cox_desc$estimate), 2),
  low = round(exp(model_cox_desc$conf.low), 2),
  high = round(exp(model_cox_desc$conf.high), 2),
  p = round(model_cox_desc$p.value, 5)
)
tab_cox$IC_95 <- paste("[", tab_cox$low, "-", tab_cox$high, "]", sep = "")

tab_cox %>%
  select(label, OR, IC_95, p) %>%
  kable(
    caption = "Cox proportional hazard model of in-ICU mortality",
    col.names = c("", "OR", "CI (95%)", "p"),
    align = c("c", "c", "c", "c")
  ) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))
```

```{r echo=FALSE, dpi=400}
tab_cox$label <- factor(tab_cox$label, levels = rev(tab_cox$label))

ggplot(tab_cox, aes(x = label, y = OR, ymin = low, ymax = high)) +
  geom_pointrange(color = "#455a64") +
  geom_hline(yintercept = 1, lty = 2, color = "#455a64") + 
  coord_flip() +
  labs(x = "", y = "OR (IC 95%)") +
  theme_bw()
```

## Kaplan-Meier curve

```{r, dpi=400}
survplot <- ggsurvplot(
  fit,
  data = cohort,
  size = 1.2,
  pval = "Hazard ratio 0.97, 95% CI [0.68-1.37]\np = 0.845",
  pval.size = 7,
  pval.coord = c(0.2, 0.10),
  risk.table = TRUE,
  risk.table.height = 0.2,
  tables.theme = theme_cleantable(),
  tables.y.text.col = FALSE,
  linetype = "strata",
  palette = palette,
  xlab = "Days since admission",
  legend = c(0.85, 0.7),
  legend.title = "",
  legend.labs = c("Normoxia", "Hyperoxia"),
  fontsize = 7,
  font.x = 20,
  font.y = 20,
  font.tickslab = 18,
  font.legend = 18,
  ggtheme = theme_classic()
)
survplot$table <- survplot$table +
  theme(plot.title = element_text(size = 20),
        axis.text.y = element_text(size = 18))

survplot
```

```{r}
ggsave(filename = "figure3_survival_curve.eps",
       plot = print(survplot),
       width = 20, height = 13, dpi = 600)
```

# Sensitivy analyses

## Different TWA-PaO2 cutoffs

```{r}
cohort <- cohort %>%
  mutate(
    hypero2_100 = ifelse(twavg_po2 > 100, 1, 0),
    hypero2_100 = factor(hypero2_100),
    hypero2_200 = ifelse(twavg_po2 > 200, 1, 0),
    hypero2_200 = factor(hypero2_200),
    hypero2_300 = ifelse(twavg_po2 > 300, 1, 0),
    hypero2_300 = factor(hypero2_300),
  )
```

```{r}
cohort %>% 
  summarise(
    across(hypero2_100:hypero2_300, ~ twoby2(1 - as.numeric(.x), 1 - as.numeric(death_icu), print = FALSE)$measures[2,])
  ) %>% 
  add_row(
    hypero2_100 = twoby2(1 - as.numeric(cohort$hypero2_100), 1 - as.numeric(cohort$death_icu), print = FALSE)$p.value[1],
    hypero2_200 = twoby2(1 - as.numeric(cohort$hypero2_200), 1 - as.numeric(cohort$death_icu), print = FALSE)$p.value[1],
    hypero2_300 = twoby2(1 - as.numeric(cohort$hypero2_300), 1 - as.numeric(cohort$death_icu), print = FALSE)$p.value[1],
  ) %>% 
  mutate(var = c("or", "or_low", "or_high", "p_value")) %>% 
  pivot_longer(starts_with("hypero2_")) %>% 
  pivot_wider(names_from = "var", values_from = "value") %>% 
  mutate_if(is.numeric, round, digits = 3) %>% 
  mutate(ic = paste0("[", or_low, "-", or_high, "]")) %>%
  select(name, or, ic, p_value) %>% 
  kable(
    caption = "Cutoffs for hyperoxemia, undadjustd analyses (ICU mortality)",
    col.names = c("", "OR", "IC (95%)", "p"),
    align = c("c", "c", "c", "c")
  ) %>% 
  kable_styling(bootstrap_options = c("striped", "hover"))
```

```{r}
model_100 <- glm(
  death_icu ~ hypero2_100 + sofa + comorbidity_liver + comorbidity_cancer +
    lactate_max_firstday + ph_min_firstday + bicarbonate_min_firstday +
    bilirubin_max_firstday + creatinine_max_firstday + source_other,
  data = cohort,
  family = "binomial"
)
model_100_desc <- tidy(model_100)

summary(model_100)
```

```{r}
table_sensitivity <- tibble(
  label = "TWA-PaO2 > 100 mmHg",
  OR = round(exp(model_100_desc$estimate[2]), 2),
  low = round(exp(confint_tidy(model_100)), 2)$conf.low[2],
  high = round(exp(confint_tidy(model_100)), 2)$conf.high[2],
  p = round(model_100_desc$p.value[2], 5)
) 
```

```{r}
model_200 <- glm(
  death_icu ~ hypero2_200 + sofa + comorbidity_liver + comorbidity_cancer +
    lactate_max_firstday + ph_min_firstday + bicarbonate_min_firstday +
    bilirubin_max_firstday + creatinine_max_firstday + source_other,
  data = cohort,
  family = "binomial"
)
model_200_desc <- tidy(model_200)

summary(model_200)
```

```{r}
table_sensitivity <- table_sensitivity %>% 
  add_row(
    label = "TWA-PaO2 > 200 mmHg",
    OR = round(exp(model_200_desc$estimate[2]), 2),
    low = round(exp(confint_tidy(model_200)), 2)$conf.low[2],
    high = round(exp(confint_tidy(model_200)), 2)$conf.high[2],
    p = round(model_200_desc$p.value[2], 5)
)
```

```{r}
model_300 <- glm(
  death_icu ~ hypero2_300 + sofa + comorbidity_liver + comorbidity_cancer +
    lactate_max_firstday + ph_min_firstday + bicarbonate_min_firstday +
    bilirubin_max_firstday + creatinine_max_firstday + source_other,
  data = cohort,
  family = "binomial"
)
model_300_desc <- tidy(model_300)

summary(model_300)
```

```{r}
table_sensitivity <- table_sensitivity %>% 
  add_row(
    label = "TWA-PaO2 > 300 mmHg",
    OR = round(exp(model_300_desc$estimate[2]), 2),
    low = round(exp(confint_tidy(model_300)), 2)$conf.low[2],
    high = round(exp(confint_tidy(model_300)), 2)$conf.high[2],
    p = round(model_300_desc$p.value[2], 5)
)
```

```{r}
table_sensitivity %>% 
  mutate(IC_95 = paste0("[", low, "-", high, "]")) %>%
  select(label, OR, IC_95, p) %>%
  kable(
    caption = "Multivariate sensitivity analyses of ICU mortality",
    col.names = c("", "OR", "IC (95%)", "p"),
    align = c("c", "c", "c", "c")
  ) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))
```
